Kolin patch, and english voice set for mercs
Took the English voice set that a bunch of mods shared (included LASS) and assigned it to mercs and ISG, and the voice set Women of the Zone had assigned for ecologists.
Patch for MERCS Kolin Model Complete. Half the LCS wearing mercs and eco mercs will use that model. I thought of patching one of the STS wearing ones instead, but wasn't sure.